requirements.txt
python-telegram-bot==20.0
Flask==2.3.3
gunicorn==20.1.0